refactor(exports/es): move test classes
